pyproject.toml
[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''


[tool.isort]
profile = "black"


[tool.poetry]
name = "assina-docx"
version = "0.1.0"
description = ""
authors = ["Victor Nunes <nunessvictorr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^1.1.2"
imgkit = "^1.2.3"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
